Migrate user routes to typescript
