+++ jest.config.js
@@ -1,10 +1,10 @@
 module.exports = {
-    roots: ['<rootDir>/test'],
-    testMatch: [
-        '**/__tests__/**/*.+(ts|tsx|js)',
-        '**/?(*.)+(spec|test).+(ts|tsx|js)',
-    ],
-    transform: {
-        '^.+\\.(ts|tsx)$': 'ts-jest',
-    },
+  roots: ['<rootDir>/test'],
+  testMatch: [
+    '**/__tests__/**/*.+(ts|tsx|js)',
+    '**/?(*.)+(spec|test).+(ts|tsx|js)',
+  ],
+  transform: {
+    '^.+\\.(ts|tsx)$': 'ts-jest',
+  },
 };
